Controller finish

+++ ObracunPlace_HibernateORM/src/tisaj/obracunplace/controller/ObradaVrstaPrimanja.java
@@ -35,7 +35,9 @@ public class ObradaVrstaPrimanja extends Obrada<VrstaPrimanja> implements Obrada
      }
     
      public void obrisi(VrstaPrimanja vp) throws ObracunPlaceException{
-            
+         if(!vp.getObracuni().isEmpty()){
+             throw  new ObracunPlaceException("Ne može se obrisati vrsta primanja jer je u obračunu");
+         }
          dao.delete(vp);
      }
